# multicore-safe version of StoreVector