See the diff for lengthy explanations of why `unsafe_{get,set}` cannot be assumed correct with OCaml 5.

The performance results are fairly noisy on my machine.

Before, three timings: 6.6s, 6.1s, 6.9s
After, three timings: 6.3s, 6.6s, 6.7s

My performance conclusion is that this is fine.
It is not like we have a choice anyway, memory-unsoundess shall not be tolerated.